requirements.txt
flask
tensorflow
numpy
scikit-learn
Pillow